requirements.txt
fastapi==0.112.0
mysql-connector-python==8.0.16
pylint==3.3.4
pytest==8.3.4
python-dotenv==1.0.1
SQLAlchemy==2.0.38
uvicorn==0.34.0
